Update to use gauge
