+++ perfect-group-allocation-backend/Cargo.toml
@@ -53,12 +53,13 @@ tokio-metrics = { version = "0.3" }
 zero-cost-templating = { git = "https://github.com/mohe2015/zero-cost-templating.git", branch = "work" }
 bytes = "1"
 
-opentelemetry = { version = "0.21" }
-opentelemetry-otlp = { version = "0.14", features = ["metrics", "http-proto", "reqwest-client", "logs"] }
-opentelemetry_sdk = { version = "0.21", features = ["rt-tokio"] }
-opentelemetry-semantic-conventions = "0.13"
-opentelemetry-appender-log = "0.2"
-opentelemetry-appender-tracing = "0.2"
+# https://github.com/open-telemetry/opentelemetry-rust/pull/1410
+opentelemetry = { git = "https://github.com/open-telemetry/opentelemetry-rust.git", features = ["metrics", "otel_unstable"] }
+opentelemetry_sdk = { git = "https://github.com/open-telemetry/opentelemetry-rust.git", features = ["metrics", "rt-tokio"] }
+opentelemetry-otlp = { git = "https://github.com/open-telemetry/opentelemetry-rust.git", features = ["metrics", "http-proto", "reqwest-client", "logs"] }
+opentelemetry-semantic-conventions = { git = "https://github.com/open-telemetry/opentelemetry-rust.git" }
+opentelemetry-appender-log = { git = "https://github.com/open-telemetry/opentelemetry-rust.git" }
+opentelemetry-appender-tracing = { git = "https://github.com/open-telemetry/opentelemetry-rust.git" }
 tracing = "0.1"
-tracing-opentelemetry = "0.22"
+tracing-opentelemetry = { git = "https://github.com/mohe2015/tracing-opentelemetry.git" }
 tracing-subscriber = "0.3"
